api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Mikołaj Data <miki3867@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.11"
uvicorn = "^0.24.0.post1"
channels = "^4.0.0"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.0.1"
pgvector = "^0.2.4"
django-stubs-ext = "^4.2.7"
requests = "^2.32.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
